pyproject.toml
[project]
name = "turtlebot3-gazebo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.136.0",
    "numpy==1.26.4",
    "open-clip-torch>=3.3.0",
    "openai>=2.32.0",
    "opencv-python>=4.11.0.86",
    "pillow>=12.2.0",
    "pip>=26.0.1",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "segment-anything>=1.0",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "transforms3d>=0.4.2",
    "uvicorn[standard]>=0.45.0",
]
